декомпозирует компонент OfferCardPage
