+++ project/src/components/offers-list/offers-list.tsx
